chore: moves e2e tests to test/fields folder
